+++ src/netcore/Wexflow.Core.Db.Firebird/Db.cs
@@ -290,7 +290,7 @@ namespace Wexflow.Core.Db.Firebird
                             LaunchType = (LaunchType)(int)reader[Entry.ColumnName_LaunchType],
                             Status = (Status)(int)reader[Entry.ColumnName_Status],
                             StatusDate = (DateTime)reader[Entry.ColumnName_StatusDate],
-                            WorkflowId = (int)(int)reader[Entry.ColumnName_WorkflowId],
+                            WorkflowId = (int)reader[Entry.ColumnName_WorkflowId],
                             JobId = (string)reader[Entry.ColumnName_JobId]
                         };
 
@@ -406,7 +406,7 @@ namespace Wexflow.Core.Db.Firebird
                             LaunchType = (LaunchType)(int)reader[Entry.ColumnName_LaunchType],
                             Status = (Status)(int)reader[Entry.ColumnName_Status],
                             StatusDate = (DateTime)reader[Entry.ColumnName_StatusDate],
-                            WorkflowId = (int)(int)reader[Entry.ColumnName_WorkflowId],
+                            WorkflowId = (int)reader[Entry.ColumnName_WorkflowId],
                             JobId = (string)reader[Entry.ColumnName_JobId]
                         };
 
@@ -467,7 +467,7 @@ namespace Wexflow.Core.Db.Firebird
                         LaunchType = (LaunchType)(int)reader[Entry.ColumnName_LaunchType],
                         Status = (Status)(int)reader[Entry.ColumnName_Status],
                         StatusDate = (DateTime)reader[Entry.ColumnName_StatusDate],
-                        WorkflowId = (int)(int)reader[Entry.ColumnName_WorkflowId],
+                        WorkflowId = (int)reader[Entry.ColumnName_WorkflowId],
                         JobId = (string)reader[Entry.ColumnName_JobId]
                     };
 
@@ -510,7 +510,7 @@ namespace Wexflow.Core.Db.Firebird
                         LaunchType = (LaunchType)(int)reader[Entry.ColumnName_LaunchType],
                         Status = (Status)(int)reader[Entry.ColumnName_Status],
                         StatusDate = (DateTime)reader[Entry.ColumnName_StatusDate],
-                        WorkflowId = (int)(int)reader[Entry.ColumnName_WorkflowId],
+                        WorkflowId = (int)reader[Entry.ColumnName_WorkflowId],
                         JobId = (string)reader[Entry.ColumnName_JobId]
                     };
 
@@ -602,7 +602,7 @@ namespace Wexflow.Core.Db.Firebird
                             LaunchType = (LaunchType)(int)reader[HistoryEntry.ColumnName_LaunchType],
                             Status = (Status)(int)reader[HistoryEntry.ColumnName_Status],
                             StatusDate = (DateTime)reader[HistoryEntry.ColumnName_StatusDate],
-                            WorkflowId = (int)(int)reader[HistoryEntry.ColumnName_WorkflowId]
+                            WorkflowId = (int)reader[HistoryEntry.ColumnName_WorkflowId]
                         };
 
                         entries.Add(entry);
@@ -647,7 +647,7 @@ namespace Wexflow.Core.Db.Firebird
                             LaunchType = (LaunchType)(int)reader[HistoryEntry.ColumnName_LaunchType],
                             Status = (Status)(int)reader[HistoryEntry.ColumnName_Status],
                             StatusDate = (DateTime)reader[HistoryEntry.ColumnName_StatusDate],
-                            WorkflowId = (int)(int)reader[HistoryEntry.ColumnName_WorkflowId]
+                            WorkflowId = (int)reader[HistoryEntry.ColumnName_WorkflowId]
                         };
 
                         entries.Add(entry);
@@ -694,7 +694,7 @@ namespace Wexflow.Core.Db.Firebird
                             LaunchType = (LaunchType)(int)reader[HistoryEntry.ColumnName_LaunchType],
                             Status = (Status)(int)reader[HistoryEntry.ColumnName_Status],
                             StatusDate = (DateTime)reader[HistoryEntry.ColumnName_StatusDate],
-                            WorkflowId = (int)(int)reader[HistoryEntry.ColumnName_WorkflowId]
+                            WorkflowId = (int)reader[HistoryEntry.ColumnName_WorkflowId]
                         };
 
                         entries.Add(entry);
@@ -807,7 +807,7 @@ namespace Wexflow.Core.Db.Firebird
                             LaunchType = (LaunchType)(int)reader[HistoryEntry.ColumnName_LaunchType],
                             Status = (Status)(int)reader[HistoryEntry.ColumnName_Status],
                             StatusDate = (DateTime)reader[HistoryEntry.ColumnName_StatusDate],
-                            WorkflowId = (int)(int)reader[HistoryEntry.ColumnName_WorkflowId]
+                            WorkflowId = (int)reader[HistoryEntry.ColumnName_WorkflowId]
                         };
 
                         entries.Add(entry);
